Update dock.tsx
Fix typo causing React warning.

```
Warning: React does not recognize the `mouseX` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `mousex` instead. If you accidentally passed it from a parent component, remove it from the DOM element.
```